+++ tests/testthat/test-ensemble.R
@@ -71,7 +71,8 @@ test_that("ensemble works", {
   expect_equal(sort(unique(sims[["summary"]][["j"]])), 1)
 
   # Check returned constants and init
-  expect_equal(sort(unique(sims[["constants"]][["summary"]][["variable"]])), c("a0", "a1", "a2"))
+  expect_equal(sort(unique(sims[["constants"]][["summary"]][["variable"]])),
+               c("a0", "a1", "a2"))
   expect_equal(sort(unique(sims[["init"]][["summary"]][["variable"]])), c(
     "Compensatory_behaviour",
     "Food_intake",